requirements.txt
# Core dependencies for data handling, tensor networks, and utilities
pandas>=2.0.0
numpy==2.2.6
quimb==1.11.0
tqdm>=4.64.0
psutil>=5.9.0
scipy>=1.9.0
networkx>=2.8.0
plotly>=5.10.0
scikit-learn>=1.1.0
matplotlib>=3.10.3
numexpr>=2.10.0

# Tensor contraction optimization
cotengra>=0.2.0
optuna>=3.0.0  # Hyper-optimizer for cotengra
cmaes>=0.10.0  # Optional hyper-optimizer
nevergrad>=1.0.0  # Optional hyper-optimizer

# GPU support (cupy for single-GPU, dask for parallelization)
cupy-cuda12x  # Matches CUDA 12.6
dask>=2025.4.0
dask-cuda>=24.0.0  # For multi-GPU on Linux; version may need adjustment